chris-atmospheric-correction/src/main/java/org/esa/beam/chris/operators/ZeroOpImage.java
/*
 * Copyright (C) 2010 Brockmann Consult GmbH (info@brockmann-consult.de)
 *
 * This program is free software; you can redistribute it and/or modify it
 * under the terms of the GNU General Public License as published by the Free
 * Software Foundation; either version 3 of the License, or (at your option)
 * any later version.
 * This program is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
 * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
 * more details.
 *
 * You should have received a copy of the GNU General Public License along
 * with this program; if not, see http://www.gnu.org/licenses/
 */
package org.esa.beam.chris.operators;

import javax.media.jai.ComponentSampleModelJAI;
import javax.media.jai.ImageLayout;
import javax.media.jai.OpImage;
import javax.media.jai.PlanarImage;
import javax.media.jai.SourcelessOpImage;
import java.awt.Rectangle;
import java.awt.image.ColorModel;
import java.awt.image.DataBuffer;
import java.awt.image.SampleModel;
import java.awt.image.WritableRaster;
import java.util.Map;

/**
 * Zero image.
 *
 * @author Ralf Quast
 * @version $Revision$ $Date$
 * @since BEAM 4.2
 */
class ZeroOpImage extends SourcelessOpImage {

    /**
     * Creates an image filled with zeros.
     *
     * @param w the image width.
     * @param h the image height.
     *
     * @return the image.
     */
    public static OpImage createImage(int w, int h) {
        final SampleModel sampleModel = new ComponentSampleModelJAI(DataBuffer.TYPE_BYTE, w, h, 1, w, new int[]{0});
        final ColorModel colorModel = PlanarImage.createColorModel(sampleModel);
        final ImageLayout imageLayout = new ImageLayout(0, 0, w, h, 0, 0, w, h, sampleModel, colorModel);

        return new ZeroOpImage(imageLayout, null, sampleModel, 0, 0, w, h);
    }

    private ZeroOpImage(ImageLayout imageLayout, Map map, SampleModel sampleModel, int minX, int minY, int w, int h) {
        super(imageLayout, map, sampleModel, minX, minY, w, h);
    }

    @Override
    protected void computeRect(PlanarImage[] planarImages, WritableRaster target, Rectangle targetRectangle) {
    }
}
